requirements.txt
PyQt6
SpeechRecognition
pyttsx3
python-dotenv
google-generativeai